stage0Module3Task5

+++ src/main/java/lang/print/gaps/task5/NumberDivider.java
@@ -3,7 +3,7 @@ package lang.print.gaps.task5;
 public class NumberDivider {
     void divide(int toBeDivided) {
         toBeDivided = 19;
-        float sum = toBeDivided % 5;
+        double sum = toBeDivided % 5;
         System.out.println(sum);
 
     }
